Train on the Ewbank -> rating transformation
Changes the transformation from Ewbank's grades to ratings so that it's
consistent with the experimental observation that natural ratings have a
linear relationship with Ewbank's grades.  The previous transform
suggested a linear relationship between the *gamma* ratings and the
Ewbank grade.

This also introduces a single hyperparameter representing a scaling
factor to apply to intervals on Ewbank's scale to transform them to
intervals on the natural rating scale.  Setting this hyperparameter to 0
means the conventional grades are ignored.