Float the reference route
Do not constrain the reference route to have rating 0.  It isn't
necessary for identifiability with the gamma priors, and it introduces a
weird bias into the model at r = 0.

Also change the way the reference route is chosen: try and find the mode
of the route grades, and align that to r = 0.