Switch from gamma to log-normal priors
On the natural ratings scale, the density of the gamma distribution
drops off sharply past the mode.  While this seemed plausible, the
resulting prior is perhaps too strong.

Instead, use a normal distribution for the natural ratings priors.
This has a similar shape on the gamma ratings scale, but allows for
climber and route ratings much greater than the mode.