Use flattened ascents responses
Pass the "flatten" parameter when retrieving ascents from theCrag's API,
instead of "thin=1&withdata=xxx".  This appears to provide better access
to fields, in particular, the gradeScore and pitch.  The dictionary
structure from the default response can be used to construct the flatten
parameter.

Having the gradeScore on the ascent makes parsing the grade config
redundant, so I've removed relevant code.

Having the pitch info allows pitches of a multipitch to be treated as
independent routes.  The current logic is fairly basic:

- if pitches are available for an ascent, each pitch is a separate route
suffixed with "Pn", otherwise the ascent is associated with the full
route.

- pitch ticks are defaulted to the overall ascent tick if unavailable

- per-pitch ascent grades aren't used for now because they're not
conveniently available as a gradeScore; the ascent grade is used for the
individual pitches too.

Also adds golden tests for data_prep_json and unit tests for some of the
additional functions in data_prep_functions.