Calculate page gaps from within Python
Replaces the "gap" column in "pages.csv" with a "timestamp" column.  The
calculation of the gaps has been moved into the whole_history_rating
module.

This reduces the preprocessing required to use the climbing_ratings
Python package; in particular, it makes it easier for the results of an
SQL query to be used (without needing analytic functions in the query to
compute pairwise differences).