Fix state leaking across tests
The WholeHistoryRating unit tests were dependent on
Climber.wiener_variance being set in the Climber unit tests.  When
testing the package as a whole, the tests still passed due to the
ordering of the files.

This is a temporary workaround; I'll stop setting the global state from
the tests as part of another refactor.