Parse styles
Adds a "style" column to the raw ascents tables.  This requires a change
in the call to the JSON endpoint, and a dummy value for logbooks.

Also updates the test data.  The JSON data is mostly derived from
raw_ascents.csv but with some manual edits for pitches and styles.

This commit does not write the style to the normalized tables, so the
output from R is still compatible with the Python script.