Encapsulate ascent and page inputs to WholeHistoryRating
Rather than passing a flat bunch of arguments to WholeHistoryRating,
organize ascent and page data into AscentTable and PageTable objects.

Also moves extract_slices into whole_history_rating and adds a unit test
for it.  I think it's better to make the interface in terms of (much
more comprehensible) relational tables and to treat the slices as an
implementation detail.