Add expand_to_slices_sparse
With ascents ordered primarily by climber and timestamp, the ascents for
a particular climber-style may not be contiguous.  Add a new function
expand_to_slices_sparse to deal with this case.  While it's possible to
use expand_to_slices_sparse everywhere expand_to_slices is used, I don't
expect it to be as efficient.

Currently the new function is only used in unit tests, but it will be
used for styles in the future.

Also changes the interface of expand_to_slices() for consistency.