Use extended precision for Bradley-Terry calculation
Use extended precision (long double) within the Bradley-Terry module, to
avoid overflow and loss of precision when working with gamma ratings.
Pass double precision natural ratings in from WholeHistoryRating, and
return the derivatives as double precision, so the change in precision
is hidden.

This has a significant (order of 10%) performance cost.