Save RDS files from data_prep
The results of the estimation script need to be combined with the
various tables in "dfs", which aren't currently written in full.
This is annoying when attempting to resume analysis because the JSON
files are slow to read, or may not even be compatible with the current
code.

To mitigate this, change the data prep scripts to save the contents of
"df_raw" and "dfs" as ".rds" files.