Make CSV parsing robust to format changes
Tolerate missing, re-ordered, and additional columns in the CSV inputs
to the run_estimation script.  Additionally, allow the style_pages.csv
file to be missing completely.

This makes the script compatible with inputs from before the
introduction of style-pages.  This is asserted with tests.

This approach also adds some forwards-compatibility with new columns
(providing they have suitable defaults).