Output the best model, rather than the last model
While iterating, keep a copy of the best model found so far, and write
the results for the best model rather than the last model update.

While usually the log-likelihood monotonically increases each iteration,
sometimes it may temporarily diverge even if it is generally
well-behaved.  This means that increasing the iteration limit will not
make the model worse (in terms of log-likelihood).

Print a warning if the model is diverging.

Adds a copy() implementation to WholeHistoryRating so that the ratings
can be snapshotted.