Increase default variance of climber and route priors
Improvements in numerical stability and warnings mean that larger
variances can be used without the model updates diverging.

Increase the default variance from 1 to 4 for both climbers and routes.
This better reflects the wide dispersion of performance and difficulty
in reality, and will most likely improve route rating estimates when the
priors are uninformed (i.e. the prior rating is 0 for all routes).

Of course, relying on the defaults is not as good as carefully-chosen
hyperparameters.  For a given dataset, priors should be chosen based on
experimentation and existing knowledge about the populations of climbers
and routes.