Drop aux parameters from get_bt_derivatives
Drop the auxiliary parameters from get_bt_summation_terms and
get_bt_derivatives in the bradley_terry module.

The auxiliary terms are always summed with the player terms, which can
be done prior to calling get_bt_derivatives.  This simplifies the API
and also paves the way for skipping processing of styles altogether.