Use alternative factorization of cosh
In the evaluation of the second derivative of the logistic function used
for each ascent in the Bradley-Terry model, use the form

  -exp(d) / (exp(d) + 1)^2

Since exp(d) is already calculated for the first derivative, this means
the second derivative can be calculated with just a few additional
arithmetic operations.  This saves additional exp() calls that are
usually part of the software implementation of cosh().

Even though the denominator is potentially a bit less stable, there
appears to be no loss of precision, and runtime is reduced by ~35%.